agregue el middleware para administrador. hice parte del crud de pedidos. Y varias cosas mas

+++ routes/web.php
@@ -3,6 +3,8 @@
 use Illuminate\Support\Facades\Route;
 
 use Illuminate\Support\Facades\Artisan;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\BlogController;
@@ -11,6 +13,8 @@ use App\Http\Controllers\CuponController;
 use App\Http\Controllers\CostoDeEnvioController;
 use App\Http\Controllers\CanastaController;
 use App\Http\Controllers\PostController;
+use App\Http\Controllers\PedidoController;
+use App\Http\Controllers\NuestrosProductosController;
 
 /*
 |--------------------------------------------------------------------------
@@ -61,10 +65,11 @@ Route::get('/nosotros', function () {
 | Esta ruta solo devuelve una vista, por lo tanto no es 
 | necesario utilizar un controlador.
 */
-Route::get('/nuestros_productos', function () {
-    //return view('nuestros_productos');
-    return "nuestros_productos";
-})->name('nuestros_productos');
+Route::get('/nuestros_productos', NuestrosProductosController::class)->name('nuestros_productos');
+/*Route::get('/nuestros_productos', function () {
+    return view('nuestros_productos');
+    //return "nuestros_productos";
+})->name('nuestros_productos');*/
 
 
 
@@ -127,13 +132,13 @@ Route::controller(BlogController::class)->group(function () {
 | de CRUD para productos. 
 */
 Route::controller(ProductoController::class)->group(function () {
-    Route::get('productos', 'index')->name('productos.index');
+    Route::get('productos', 'index')->name('productos.index')->middleware('acceso.administrador');
     /*Route::get('productos/create', 'create')->name('productos.create')->middleware('administrador');*/
-    Route::post('productos', 'store')->name('productos.store');
+    Route::post('productos', 'store')->name('productos.store')->middleware('acceso.administrador');
     /*Route::get('productos/{producto}', 'show')->name('productos.show')->middleware('administrador');*/
-    Route::get('productos/{producto}/edit', 'edit')->name('productos.edit');
-    Route::put('productos/{producto}', 'update')->name('productos.update');
-    Route::delete('productos/{producto}', 'destroy')->name('productos.destroy');
+    Route::get('productos/{producto}/edit', 'edit')->name('productos.edit')->middleware('acceso.administrador');
+    Route::put('productos/{producto}', 'update')->name('productos.update')->middleware('acceso.administrador');
+    Route::delete('productos/{producto}', 'destroy')->name('productos.destroy')->middleware('acceso.administrador');
 });
 
 
@@ -148,13 +153,13 @@ Route::controller(ProductoController::class)->group(function () {
 | de CRUD para cupones. 
 */
 Route::controller(CuponController::class)->group(function () {
-    Route::get('cupones', 'index')->name('cupones.index');
+    Route::get('cupones', 'index')->name('cupones.index')->middleware('acceso.administrador');
     /*Route::get('cupones/create', 'create')->name('cupones.create')->middleware('administrador');*/
-    Route::post('cupones', 'store')->name('cupones.store');
+    Route::post('cupones', 'store')->name('cupones.store')->middleware('acceso.administrador');
     /*Route::get('cupones/{cupon}', 'show')->name('cupones.show')->middleware('administrador');*/
-    Route::get('cupones/{cupon}/edit', 'edit')->name('cupones.edit');
-    Route::put('cupones/{cupon}', 'update')->name('cupones.update');
-    Route::delete('cupones/{cupon}', 'destroy')->name('cupones.destroy');
+    Route::get('cupones/{cupon}/edit', 'edit')->name('cupones.edit')->middleware('acceso.administrador');
+    Route::put('cupones/{cupon}', 'update')->name('cupones.update')->middleware('acceso.administrador');
+    Route::delete('cupones/{cupon}', 'destroy')->name('cupones.destroy')->middleware('acceso.administrador');
 });
 
 
@@ -169,13 +174,13 @@ Route::controller(CuponController::class)->group(function () {
 | de CRUD para costos_de_envio. 
 */
 Route::controller(CostoDeEnvioController::class)->group(function () {
-    Route::get('costos_de_envio', 'index')->name('costos_de_envio.index');
+    Route::get('costos_de_envio', 'index')->name('costos_de_envio.index')->middleware('acceso.administrador');
     /*Route::get('costos_de_envio/create', 'create')->name('costos_de_envio.create')->middleware('administrador');*/
-    Route::post('costos_de_envio', 'store')->name('costos_de_envio.store');
+    Route::post('costos_de_envio', 'store')->name('costos_de_envio.store')->middleware('acceso.administrador');
     /*Route::get('costos_de_envio/{costo_de_envio}', 'show')->name('costos_de_envio.show')->middleware('administrador');*/
-    Route::get('costos_de_envio/{costo_de_envio}/edit', 'edit')->name('costos_de_envio.edit');
-    Route::put('costos_de_envio/{costo_de_envio}', 'update')->name('costos_de_envio.update');
-    Route::delete('costos_de_envio/{costo_de_envio}', 'destroy')->name('costos_de_envio.destroy');
+    Route::get('costos_de_envio/{costo_de_envio}/edit', 'edit')->name('costos_de_envio.edit')->middleware('acceso.administrador');
+    Route::put('costos_de_envio/{costo_de_envio}', 'update')->name('costos_de_envio.update')->middleware('acceso.administrador');
+    Route::delete('costos_de_envio/{costo_de_envio}', 'destroy')->name('costos_de_envio.destroy')->middleware('acceso.administrador');
 });
 
 
@@ -190,13 +195,13 @@ Route::controller(CostoDeEnvioController::class)->group(function () {
 | de CRUD para canastas. 
 */
 Route::controller(CanastaController::class)->group(function () {
-    Route::get('canastas', 'index')->name('canastas.index');
+    Route::get('canastas', 'index')->name('canastas.index')->middleware('acceso.administrador');
     /*Route::get('canastas/create', 'create')->name('canastas.create')->middleware('administrador');*/
-    Route::post('canastas', 'store')->name('canastas.store');
+    Route::post('canastas', 'store')->name('canastas.store')->middleware('acceso.administrador');
     /*Route::get('canastas/{canasta}', 'show')->name('canastas.show')->middleware('administrador');*/
-    Route::get('canastas/{canasta}/edit', 'edit')->name('canastas.edit');
-    Route::put('canastas/{canasta}', 'update')->name('canastas.update');
-    Route::delete('canastas/{canasta}', 'destroy')->name('canastas.destroy');
+    Route::get('canastas/{canasta}/edit', 'edit')->name('canastas.edit')->middleware('acceso.administrador');
+    Route::put('canastas/{canasta}', 'update')->name('canastas.update')->middleware('acceso.administrador');
+    Route::delete('canastas/{canasta}', 'destroy')->name('canastas.destroy')->middleware('acceso.administrador');
 });
 
 
@@ -211,13 +216,34 @@ Route::controller(CanastaController::class)->group(function () {
 | de CRUD para posts. 
 */
 Route::controller(PostController::class)->group(function () {
-    Route::get('posts', 'index')->name('posts.index');
-    /*Route::get('posts/create', 'create')->name('posts.create')->middleware('administrador');*/
-    Route::post('posts', 'store')->name('posts.store');
-    /*Route::get('posts/{post}', 'show')->name('posts.show')->middleware('administrador');*/
-    Route::get('posts/{post}/edit', 'edit')->name('posts.edit');
-    Route::put('posts/{post}', 'update')->name('posts.update');
-    Route::delete('posts/{post}', 'destroy')->name('posts.destroy');
+    Route::get('posts', 'index')->name('posts.index')->middleware('acceso.administrador');
+    /*Route::get('posts/create', 'create')->name('posts.create');*/
+    Route::post('posts', 'store')->name('posts.store')->middleware('acceso.administrador');
+    /*Route::get('posts/{post}', 'show')->name('posts.show');*/
+    Route::get('posts/{post}/edit', 'edit')->name('posts.edit')->middleware('acceso.administrador');
+    Route::put('posts/{post}', 'update')->name('posts.update')->middleware('acceso.administrador');
+    Route::delete('posts/{post}', 'destroy')->name('posts.destroy')->middleware('acceso.administrador');
+});
+
+
+
+
+/*
+|--------------------------------------------------------------------------
+| pedidos
+|--------------------------------------------------------------------------
+| La ruta de pedidos es administrada por el controlador 
+| PedidoController, ya que debe cumplir con la funciones 
+| de CRUD para pedidos. 
+*/
+Route::controller(PedidoController::class)->group(function () {
+    Route::get('pedidos', 'index')->name('pedidos.index')->middleware('acceso.administrador');
+    /*Route::get('pedidos/create', 'create')->name('pedidos.create');*/
+    Route::post('pedidos', 'store')->name('pedidos.store')->middleware('acceso.administrador');
+    /*Route::get('pedidos/{pedido}', 'show')->name('pedidos.show')->middleware('acceso.administrador');*/
+    Route::get('pedidos/{pedido}/edit', 'edit')->name('pedidos.edit')->middleware('acceso.administrador');
+    Route::put('pedidos/{pedido}', 'update')->name('pedidos.update')->middleware('acceso.administrador');
+    Route::delete('pedidos/{pedido}', 'destroy')->name('pedidos.destroy')->middleware('acceso.administrador');
 });
 
 
@@ -233,7 +259,7 @@ Route::controller(PostController::class)->group(function () {
 Route::get('/mi_perfil', function () {
     return view('mi_perfil');
     //return "mi_perfil";
-})->name('mi_perfil');
+})->name('mi_perfil')->middleware('auth');
 
 
 
@@ -296,10 +322,10 @@ Route::get('/storage_link', function () {
 | Esta ruta solo devuelve una vista, por lo tanto no es 
 | necesario utilizar un controlador.
 */
-Route::get('/welcome', function () {
+/*Route::get('/welcome', function () {
     return view('welcome');
     //return "welcome";
-})->name('welcome');
+})->name('welcome');*/
 
 
 /*
@@ -312,4 +338,21 @@ Route::get('/welcome', function () {
 Route::get('/dashboard', function () {
     return view('dashboard');
     //return "dashboard";
-})->name('dashboard');
+})->name('dashboard')->middleware('acceso.administrador');
+
+
+
+
+/*
+|--------------------------------------------------------------------------
+| cerrar_sesion
+|--------------------------------------------------------------------------
+| Esta ruta no devuelve ninguna vista. 
+| Simplemente, ejecuta las acciones para cerrar la sesión
+*/
+Route::get('/cerrar_sesion', function (Request $request) {
+    Auth::guard('web')->logout();
+    $request->session()->invalidate();
+    $request->session()->regenerateToken();
+    return redirect('/');
+})->middleware('auth')->name('cerrar_sesion');
